+++ tools/testing/selftests/bpf/progs/test_map_ops.c
@@ -27,7 +27,7 @@ struct {
 } array_map SEC(".maps");
 
 const volatile pid_t pid;
-long err = 0;
+int err = 0;
 
 static u64 callback(u64 map, u64 key, u64 val, u64 ctx, u64 flags)
 {
