+++ tools/testing/selftests/bpf/progs/test_ringbuf_map_key.c
@@ -29,7 +29,7 @@ struct {
 int pid = 0;
 
 /* inner state */
-long seq = 0;
+long long seq = 0;
 
 SEC("fentry/" SYS_PREFIX "sys_getpgid")
 int test_ringbuf_mem_map_key(void *ctx)
