+++ tools/testing/selftests/bpf/progs/test_static_linked2.c
@@ -9,7 +9,7 @@ static volatile int static_var1 = 5;
 static volatile int static_var2 = 6;
 int var2 = -1;
 /* 8-byte aligned .rodata */
-const volatile long rovar2;
+const volatile long long rovar2;
 
 /* same "subprog" name in both files */
 static __noinline int subprog(int x)
